refactor block reader, add receipt reader, forge reads receipts
